requirements.txt
asgiref==3.2.3
Django==3.0.2
django-crispy-forms==1.8.1
django-geojson==2.12.0
django-leaflet==0.26.0
django-location-field==2.1.0
geographiclib==1.50
geopy==1.20.0
Pillow==8.1.0
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
